patch: filename and public function name shuffle.
I'm thinking of trying to loosely (loosely!) standardize the idea of
some phases for features like this one: parse, compile, eval, report;
these seem to be somewhat recurrent.

(Conventionalizing function names could actually save headaches for
the package author, because it removes some decisions that must be made
but are not interesting.  It can also help users, for obvious reasons.)